OpenWrt: 042-v5.5-arm64-csum-Fix-pathological-zero-length-calls.patch
"arm64: csum: Fix pathological zero-length calls

In validating the checksumming results of the new routine, I sadly
neglected to test its not-checksumming results. Thus it slipped through
that the one case where @buff is already dword-aligned and @len = 0
manages to defeat the tail-masking logic and behave as if @len = 8.
For a zero length it doesn't make much sense to deference @buff anyway,
so just add an early return (which has essentially zero impact on
performance).

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Will Deacon <will@kernel.org>"

Change-Id: Ie5f9876e523e0d9290860afbf9437e3b1cdea108
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>