OpenWrt: 747-v5.5-net-dsa-mv88e6xxx-Add-support-for-port-mirroring.patch
net: dsa: mv88e6xxx: Add support for port mirroring

Add support for configuring port mirroring through the cls_matchall
classifier. We do a full ingress and/or egress capture towards a
capture port. It allows setting a different capture port for ingress
and egress traffic.

It keeps track of the mirrored ports and the destination ports to
prevent changes to the capture port while other ports are being
mirrored.

Signed-off-by: Iwan R Timmer <irtimmer@gmail.com>
Reviewed-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>

Change-Id: Ibe3e92f457d0b440ba2e59a520e518c55c929602
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>