arm64: dts: ipq807x: Configure GPIO30 to high
Default value of GPIO_30 is low. This change will configure
GPIO_30 to high to power on SDX core on HK01.c5 board.

If GPIO_30 is not high, then SDX core will not get powered
on.

This TLMM GPIO is used on hk01.c5 boards in which PMIC GPIOs
are not present for powering on the SDX core during each
reset. Note that SDX will not remain powered on during warm
reset when we use TLMM GPIO.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: I53200ab5ab4c79bb3c13360268e7240538ef9555