drivers: bus: mhi: Update MHITest driver
mhi_alloc_controller and mhi_free_controller APIs are added in MHI
core recently and so the same can be made use of in mhitest driver. This
patch handles this change accordingly.

mhi_unregister_controller was getting called after mhitest_global_soc_reset
which is not correct, therefore this order was updated.

Also, mhi controller was not free'd in rmmod path leading to mem leak,
so the same has been fixed now.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: Id2c6f167188a408c74118689186bcedfdb256aa8