drivers: bus: mhi: Allocate mhi device for sdx
The sdx controller needs to allocate memory for the controller
and the device. The core MHI driver API mhi_alloc_controller
only allocates memory for the controller.

Therefore, this patch adds a new API which can allocate memory
for both controller and the device. This is a temporary patch
which must be fixed and the core APIs must be used in future
for all such kinds of allocation.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: If77705471436fbbfe8f61a641e162885ed78ea32