OpenWrt: 447-mtd-spinand-gigadevice-Add-support-for-GD5F4GQ4xC.patch
mtd: spinand: gigadevice: Add support for
 GD5F4GQ4xC

This adds support for the following 4GiB chips:
GD5F4GQ4RCYIG 1.8V
GD5F4GQ4UCYIG 3.3V

The datasheet can be found here:
https://www.novitronic.ch/sixcms/media.php/2/DS-00173-GD5F4GQ4xCxIG-Rev1.574695.pdf

The GD5F4GQ4UCYIGT (3.3V) version is used on the Imagination
Technologies Creator Ci40 (Marduk), the 1.8V version was not tested.

This device only works in single SPI mode and not in dual or quad mode
for me on this board.
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

Change-Id: I3e7fc30418d23e70e3afd8e62bcc1f3e5be842e9
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>