OpenWrt: Deleted 553-ubifs-Add-option-to-create-UBI-FS-version-4-on-empty.patch
kernel: Create UBIFS v5 volume
In commit 87b35c16adcf ("kernel: ubifs: create use file system format 4
by default") we changed the default UBIFS version used when the kernel
creates a new volume from 5 to 4. UBIFS v5 was added in kernel 4.10 and
only kernel 4.10 and later can read it.
We changed the kernel to create version 4 volumes also on more recent
kernel versions to make it possible to downgrade to an OpenWrt version
with kernel 4.9, the upgrade still works. This is probably not needed
any more and we can remove this patch.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

Change-Id: Ibde38aaf3b7f1f02087ffda087adcc5389325cc4
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>