OpenWrt: 760-net-dsa-mv88e6xxx-fix-vlan-setup.patch
"net: dsa: mv88e6xxx: fix vlan setup

Provide an option that drivers can set to indicate they want to receive
vlan configuration even when vlan filtering is disabled. This is safe
for Marvell DSA bridges, which do not look up ingress traffic in the
VTU if the port is in 8021Q disabled state. Whether this change is
suitable for all DSA bridges is not known.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>
Signed-off-by: DENG Qingfang <dqfext@gmail.com>"

Change-Id: Ife4a14328f6fd7c45db623fc843916396355856b
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>