OpenWrt: 991-platform-x86-pcengines-apuv2-revert-simswitch.patch
Revert "platform/x86: pcengines-apuv2: wire up simswitch gpio
 as led"

This reverts commit 5037d4ddda31c2dbbb018109655f61054b1756dc.

Commit message from linux:
The APU3+ boards have two SIM sockets, while only one of them
can be routed to the mpcie slots at a time. Selection is done
via simswap gpio.

We currently don't have a fitting subsystem for those cases yet,
so just wire it up to a LED for the time being. While this isn't
really semantically correct, it's a good compromise.

Explanation why this does not work:
This change connects the simswap to the LED subsystem of the kernel.
From my point of view, it's nonsense. If we do it this way, then this
can be switched relatively easily via the LED subsystem (trigger:
none/default-on) and that is dangerous! If this is used, it would be
unfavorable, since there is also another trigger (trigger:
heartbeat/netdev).
This LED also appears in the LuCI and can therefore be switched by the
user.

Therefore, this simswap GPIO should remain in the GPIO
subsystem and be switched via it and not be connected to the LED
subsystem. To avoid the problems mentioned above. The LED subsystem is
not made for this and it is not a good compromise, but rather dangerous.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>

Change-Id: I2fa98a069897904b0be7afe51165f6ed6798c9bf
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>