Revert "xfrm: remove get_mtu indirection from xfrm_type"
This reverts commit c7b37c769d2a5e711106a3c793140a4f46768e04.

This patch moved the logic for get_mtu from
the transform modules(esp4/esp6) to a new common function
xfrm_get_mtu(). This new function is using x->data as the
crypto aead.
The data field of the struct xfrm_state object is meant
to be private to the transformer and hence should only
be interpreted by the xfrm_type methods, and should not
be used in the core xfrm code.

Signed-off-by: Neelansh Mittal <neelanshm@codeaurora.org>
Change-Id: Iaa8f93f1d311f21fc2d62a0609b1eb4b0bc3ef3a