net: qrtr: Release distant nodes along the bridge node
Distant QRTR nodes can be accessed via an other node that acts as
a bridge. When a QRTR endpoint associated with a bridge node is
released, all the linked distant nodes should also be released.

This patch updates the src->sq_node with iter.index to 
release all the nodes accessible through the endpoint.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Rajkumar Ayyasamy <arajkuma@codeaurora.org>
Change-Id: I2153a115e9cf7ab4779c235225007e12036b76ec