net: cnss2: Reset driver_state only if subsystem_put is done
Currently in cnss_subsystem_put, driver_state is reset to 0 even
if subsystem_put_in_progress is true and subsystem_put is not called
from here. This causes an issue in the below scenario where rmmod
is done before VAPs are deleted and driver_state is incorrectly reset to
0. This causes the shutdown of the target to get ignored and target
remains powered on after rmmod. During the next insmod, target power on
fails as it is already powered on.

rmmod if wifi_3_0.ko
	cnss_wlan_unregister_driver
		subsystem_put (in_progress is set to true)
			SUBSYS_BEFORE_SHUTDOWN
				ol_ath_pci_remove
					cnss_subsystem_put
						subsystem_put is skipped
						driver_state reset to 0
			subsystem_shutdown
				cnss_subsys_shutdown
					if (driver_state == 0) return;
To fix the above scenario, reset the driver_state in cnss_subsystem_put
only if subsystem_put_in_progress is not set and subsystem_put is called
from this function.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: I1b792ef673dafa579c6a746bc8a5d64f5445e3a7