bus: mhi: controller: align no of irqs to power of two
SDX controller driver is requesting for 11 irqs. But core
drivers changes it to 16 to align with power of two and
allocates 16 irqs.
In such cases, the controller driver only frees the 11 irqs
which leads to a resource leak.

This patch aligns the number of irqs to power of two.
So all the irqs will be cleaned up during remove.

Signed-off-by: Rajkumar Ayyasamy <arajkuma@codeaurora.org>
Change-Id: Ic3836ce396078037f40d38f954b5eb5a11b28da7