netfilter: fix source nat issue by removing the confirm check on nat setup
When a multicast flow lands in LAN side bridge, if bridge nf call iptables
is enabled, the flow could be sent to both WAN and some of the bridge slave
interfaces in some cases. This could lead to no source nat action because
the flow is already in a confirmed state which was just set by the multicast
packet taking bridging path.
This change removes the conntrack confirm check on nat setup.

Change-Id: I38d61a403e9484e36e2e28bf6830538921556c55
Signed-off-by: Ken Zhu <guigenz@codeaurora.org>