net: cnss2: Do not shutdown target after ramdump if recovery disabled
Currently, once ramdump for PCI target is collected in
cnss_pci_dev_ramdump, target is shutdown here to prepare for next power
on after recovery.

If ramdump_enabled flag is set to 1, ramdump of PCI target is
force-collected even if recovery is disabled. In this flow, target
should not be shutdown as there might be pending IRQs from wifi driver
which would not be freed yet and disabling MSI as part of shutdown would
cause the kernel to panic.

Shutdown target after ramdump collection only if target recovery is
enabled to avoid disabling MSI when there are pending IRQs.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: Ic26cdead264a7bd0b586f0dcd5741bccffde5c4c