net: cnss2: Add config option for MSI interrupts
Add an option to configure assignment of MSI interrupts to CE,
DP and MHI using a module param. This can be used to change the MSI
configuration during runtime and best possible configuration for
a particular number of MSI's can be configured and can be choosed
based on module param. Also added msi entries for QCN9224.

Change-Id: If6d24258391c9efd8de3d372d4950a2682f46a47
Signed-off-by: Nandha Kishore Easwaran <nandhaki@codeaurora.org>