soc: qcom: ramdump: fix the use after free scenario
In _do_ramdump function we are waiting for the event completion after the
q6mem file is copied to the userspace. Event completion is sent from both
ramdump_read and ramdump_release function.

In some scenarios, ramdump_read function sends the event completion then
followed by ramdump_release function. In the in-between time, ramdump device
structure gets freed in destroy_ramdump_device.

Since the ramdump_release has the refernce to the ramdump device structure
memory, and if the memory is allocated to other object, issue is seen.

 -> crashdump_init
   -> do_elf_ramdump
     -> _do_ramdump
					-> user space process is waken up to read q6mem file
         -> wait_for_completion_timeout
					-> ramdump_read sends the completion event
 -> destroy_ramdump_device
   -> ramdump_device is freed
					-> ramdump_release is called
					Acessing the ramdump_device, if the memory
					is allocated to other object, crash is seen

To avoid this, send the completion event only from the ramdump_release function.

Change-Id: I10e1dfcb9e9c0e975dd7dc99349d3f5fa75abf9c
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>