drivers: misc: qseecom: Use reserved memory node
Previously we were hardcoding the reserved memory
region that will be used for tzapp. Since, currently
in all platforms we have a separate region "tzapp"
reserved, we can directly use the reserved memory
node reference in qseecom node and use that.

If reserved memory node reference is not given in
the node, then TZApp features will not work as the
App Notify cannot be done.

Change-id: If2275093efb31d5e9bf57df9778a8dfdee3776da
Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>