arm64: dts: ipq5018: update the mhitest node path
mhitest node is placed outside of the soc block in dtsi file, since it
doesn't have the reg property. But in board DTS files, it is enabled within
the soc block, causing the node not enabled.

Fix the mhitest node path accordingly. While at it, enable the same
for MP03.1 RDP as well.

Change-Id: I27593d16fa87c063da2bf87983bc861fb4d3df26
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>