arm: change textofs for IPQ6018 256MB profile
For 256MB profile, textofs is defined as 0x120_8000 and based on this DDR
base address is calculated.

DDR Base address = Load Address - Text Offset
		= 0x4120_8000 - 0x120_8000
		= 0x4000_0000

with the change 6a6eba152a98, the textofs is changed to 0x8000 instead
of 0x20_8000. So we need to change the textofs as 0x10_8000 to correctly
identify the DDR base address as 0x4000_0000.

Change-Id: Ibf09f1eb6d81cdf8ba342ba58626bcfba6a27d6d
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>