arm64: dts: ipq5018: move the restart reason node outside of SoC node
Restart reason node doesn't have the register property. So move it out
of the SoC node so that U-boot can patch the node properly based on the
use cases.

Also rename the dload_warm_status property to dload_warm_reset, whereas
the latter one is the correct name of the property.

Change-Id: Ibbf98b4a84bc6405aca793d8b03faaeb6a4f733c
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>