phy: qca-pcie-uniphy: fix PHY 2nd lane init
As the phy's dts node name did not have x2 in it the 2nd lane init
code did not execute ever. Default reset values for 2nd lane worked
properly so this issue was not seen till SSC mode is used.

While at it, update the documentation for the 'x2' property.

Reference Commit:
6e11fab20c5c ("phy: qca-pcie-uniphy: Fix PHY 2nd lane init").

Change-Id: Ib5a58cd524d97f9db11a7bce73758c4e0ecea2b1
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>