watchdog: qcom: updated the wdt clk rate and max timeout value
ipq8074 watchdog timer feed with 32764.5 Hz source clock and its support
maximum timeout of 32 seconds. Updated the same in the dts and driver file.
Following calculation (max_timeout = 0x10000000 / clk_rate) will be used to
derive the max WDT timeout for the akronite platform. To the remaining
platforms, "max-timeout-sec" property support was added in dts to get the
max WDT timeout.

So, To make it generic based on the device compatible, added another variable
called max_tick_count into the device data, which defines max counter value
of the WDT controller. Using this, max-timeout will be calculated in runtime
for various WDT contorllers. While at it, "max-timeout-sec" property is not
needed anymore, so removed the same.

Change-Id: Id84cd43268fff61176890e591f49572a11d9a1a9
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>