drivers: mtd: qcom_nandc: Add switch layout support
In ipq9574 platform, PBL doesn't have 4K nand support.

In order to support 4K QSPI nand in such a platform,
we have introduced this layout switch mechanism. In this,
the SBL will be in 2K and all other partitions would be
in 4K. Therefore, this change is required to make sure
during sysupgrade the layout is switched as needed
accordingly for SBL and for other partitions.

This change was made by referring to below change:
Ie9c2698a1b0320c7d83184742d4cfbb7ed1ea50f

Signed-off-by: Selvam Sathappan Periakaruppan <quic_speriaka@quicinc.com>
Signed-off-by: Rajkumar Ayyasamy <quic_arajkuma@quicinc.com>
Change-Id: Ic166b2b4c900339167b13660ce815a76b10b6a35