regulator: qcom: smd: Add support to define the bootup voltage
Kernel does not know the initial voltage set by the bootloaders.
During regulator registration, the voltage variable is just decleared
and it is zero. Based on it, regulator framework considers current
voltage as zero and trying to bringing up each regulator to minimum
supported voltage.

This introduces a dip in voltage during kernel boot and getting
stabilized once the voltage scaling comes in to picture.

To avoid the voltage dip, this patch adds support to define the
bootup voltage set by the boodloaders and based on it, regulator
framework understands that proper voltage is already set

Change-Id: Ica9dbd8ff400f1fd83b9b29b829d77fbf34b7fe7
Signed-off-by: Praveenkumar I <quic_ipkumar@quicinc.com>