firmware: qcom_scm: provide size argument to fuse blow TZ call
Certain versions of the TZ, like the one used for the IPQ50xx SoC,
expects the size of the sec.dat binary for blowing the fuses.

Add support for the same by introducing the property,
"fuse-blow-size-required". If this property is present, send the size
of the sec.dat binary as well, to the TZ. Else, fuse blow will fail.

Reference Commit:
c232fe29a37d ("drivers: qcom_scm: Provide size argument to fuse blow TZ call")

Change-Id: I85a5fdd516a9ae0dab16d89a9f4857db02d61520
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>