cpufreq: qti: Enable cpufreq for ipq60xx
IPQ60xx have different OPPs available for the CPU based on
SoC variant. This can be determined through use of an eFuse
register present in the silicon.

Added support for ipq60xx on nvmem driver which helps to
determine OPPs at runtime based on the eFuse register which
has the CPU frequency limits. opp-supported-hw dt binding
can be used to indicate the availabe OPPs for each limit.

nvmem driver also creates the "cpufreq-dt" platform_device after
passing the version matching data to the OPP framework so that the
cpufreq-dt handles the actual cpufreq implementation.

IPQ6000 variant is limited to 1.2GHz. We don't have a way to identify
this from eFuse, since only one bit is available in the eFuse. So
handled this variant explicitly by setting the corresponding bitmask.

Change-Id: Ib953ed97445c7aea31d5e49faf6dd23101c570ba
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>