coresight: tmc-etr: use the newly allocated buffer for each session
In the current design, once the buffer is allocated for the ETR, same buffer
is used across the different out_mode. For the out_mode like q6mem_stream
this approach is causing the issues.

For the q6mem_stream, we use the ETR SG mechanism, with the data page size
as 8K, where as the default data page size is 4K. If the existing buffer is
used for the q6mem_stream mode, when accessing the second page causing the
issue.

Lets fix it by using the newly allocated buffer for every session and free
the existing one. While at it, make sure all the work queued in the workqueue
is flushed when disabling the ETR.

Change-Id: Ieab0092c4314af9215192c45589b585094ac51f3
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>