soc: qcom: tzlog: Fix encrypted_tz_log support
This patch fixes the encrypted_tz_log_support
feature.

The copy_len was not populated properly, also
the fn_id's used for communication with QSEE
was incorrect and this patch fixes the same.

Without this patch, we will face errors, traces
when we try to dump the encrypted_tz_log.

Change-Id: I8cf3dd7938673e0999166457a7ff92a83c3bac0f
Signed-off-by: Selvam Sathappan Periakaruppan <quic_speriaka@quicinc.com>