drivers: usb: host: Handle TRB Host Controller event
Whenever we get the TRB Host Controller event, the
USB Event ring is full. We were seeing the event ring
full error in ipq9574 platform. This patch handles
the TRB HC Event and dequeue's the TRBs whichever
has finished so that we can avoid the event ring
full error condition.

Change-Id: I07a9d3fcb5c4ea778ff558bce2a44fc78fff88a1
Signed-off-by: Selvam Sathappan Periakaruppan <quic_speriaka@quicinc.com>